Makefile: Add Go test for schema-go
And run them from Travis.

Also replace some old Travis test commands with a single call to 'make
test'.  Now we die on the first error (e.g. we don't run the Go tests
if lint fails), but I don't think that's a big problem, and this way
we don't have to maintain two parallel lists of test targets.

Signed-off-by: W. Trevor King <wking@tremily.us>