config-linux: Forbid the empty string for mountLabel
minLength is documented in [1].  The spec is currently not clear about
how values for this property should be used, and after this commit it
is still not clear.  But the Linux.MountLabel property is not a
*string, so distinguishing between "unset" and "set to the empty
string" would be awkward in Go.  I'm not familiar enough with the
backing kernel API to be able to put RFC 2119 teeth into how the value
should be used, but I'm pretty sure we either want this commit
(forbidding the empty string) or a *string in the Go type.

[1]: https://tools.ietf.org/html/draft-wright-json-schema-validation-01#section-6.7

Signed-off-by: W. Trevor King <wking@tremily.us>