config: RFC 2119 language for uid, gid, and additionalGids
Also link to POSIX for the associated definitions.

This adds runtime-oriented requirements similar to what runtime-tools
is already checking [1].  However, runtime-tools is currently allowing
additional groups in the getgroups() response beyond those specified
in additionalGids, with "superset" wording documented in
opencontainers/runtime-tools@0ad3c80a (Fix validation for additional
groups, 2015-10-16).  I'm not sure if that superset logic is intended
to be an open door, or if it is a workaround for the
implementation-defined inclusion of the effective group ID in the
getgroups() response [2]:

  It is implementation-defined whether getgroups() also returns the
  effective group ID in the grouplist array.

  ...

  Duplication may exist, but the application needs to call getegid()
  to be sure of getting all of the information.

The wording I'm using glosses over a potential distinction between the
real and effective group ID, since it's using the gid value as the
effective group ID in the additionalGids comparison and as the real
group ID in the gid comparison.  That could conceivably be an issue if
process.args points to a setgid binary, but I think that is unlikely
to be a major issue in the wild, and it is certainly avoidable within
runtime-tools' compliance testing, so I haven't addressed it in this
commit.

[1]: https://github.com/opencontainers/runtime-tools/blob/f49d861b69e359dc9278fcfa6948034a4af58705/cmd/runtimetest/main.go#L100-L123
[2]: http://pubs.opengroup.org/onlinepubs/9699919799/functions/getgroups.html

Signed-off-by: W. Trevor King <wking@tremily.us>