runtime: Clarify ociVersion as based on the state schema
The old wording was ambiguous.  For example, if the configuration had
ociVersion set to 1.0.0 and the container was created with a 1.1.0
runtime, which version should show up in the state?

With this commit, we use the version which matches the state schema,
because the config/runtime versions used for creation don't seem
particularly important once the container has been created, while the
state schema version is important for state consumers.  For example,
if new properties were added to the state spec between 1.0.0 and
1.1.0, a consumer would want to see 1.1.0 in the state's ociVersion so
it could decide whether it could rely on those new properties.

Signed-off-by: W. Trevor King <wking@tremily.us>