config-linux: RFC 2119 MUST for absolute linux.namespaces[].path
The old language is from 72cbff67 (config-linux.md: clearly require
absolute path for namespace, 2017-03-10, #720), but without RFC 2119
language in the absolute path wording, it's not a compliance
requirement (per spec.md's "compliant" definition).  This commit
adjusts the language to bring it in line with our current wording for
maskedPaths and readonlyPaths, which we've had since 25f44dd0 (
config-linux: fix format and definitely require value of masked and
readonly paths, 2016-09-30, #587).

Signed-off-by: W. Trevor King <wking@tremily.us>