config: Dedent root paragraphs, since they aren't a list entry
I'd broken this in 2022e3ec (config: Bump Hyper-V condition from
root.path to root itself, 2017-05-18, #838), where I'd misread the
initial bolding ** as a list bullet or something.  The extra indents
don't cause a problem for the first paragraph, but they cause the "On
all other platforms..." paragraph to be interpreted as a <pre> block.

Signed-off-by: W. Trevor King <wking@tremily.us>