meeting: Bump January meeting from the 3rd to the 10th
Catching up with:

On Thu, Dec 07, 2017 at 09:14:32AM -0600, Chris Aniszczyk wrote [1]:
> btw, I bumped the January meeting a week later due to the holidays,
> that should solve issues and updated the calendar invite

I've used RDATE [2] to add the new date and EXDATE [3] to remove the
old date.  The ICS was validated with [4].

Also add some wiggle room to the README text, although I don't know
that it's worth adding text for each exception there.  After the
meeting, we can leave the exception information in the iCalendar entry
without bothering anyone, but we'd need more churn to remove it from
the README.

[1]: https://groups.google.com/a/opencontainers.org/d/msg/dev/FzclHQsMH_c/QKJMybRkBAAJ
     Subject: Re: 2017-12-06 agenda
     Date: Thu, 7 Dec 2017 09:14:32 -0600
     Message-ID: <CAJg1wMQFVvT+UiJ1gj3m2AzLQEus82eoWRVVtBEe+ZBMxskq=w@mail.gmail.com>
[2]: https://tools.ietf.org/html/rfc5545#section-3.8.5.2
[3]: https://tools.ietf.org/html/rfc5545#section-3.8.5.1
[4]: https://icalendar.org/validator.html

Signed-off-by: W. Trevor King <wking@tremily.us>