config-vm: Recycle the hook-process schema
We already have two ways to specify a process to launch (for the
container process and for hooks).  This commit recycles the
hook-process schema for launching the hypervisor.  I prefer using the
more-fundamental container process schema, but Michael felt it had too
many properties that wouldn't apply to the hypervisor-launching case
[1].

[1]: https://github.com/opencontainers/runtime-spec/pull/949#issuecomment-373730290

Signed-off-by: W. Trevor King <wking@tremily.us>