uidMappings: change order of fields for clarity
"man 7 user_namespaces" explains the format of uid_map and gid_map:
    <containerID> <hostID> <mapSize>

The order of map entries in JSON does not matter. But for the clarity of
the spec, I find it easier to understand if the order of the JSON fields is
the same as the order of the fields in the underlying uid_map/gid_map
files.

I am about to file a PR in runtime-tools because the fields in
uid_map/gid_map were parsed in the wrong order.

Signed-off-by: Alban Crequy <alban@kinvolk.io>