spec: add field for NIS domain name
A common complaint is that there is no way to explicitly set the NIS
domain name of a container (through setdomainname(2) and
kernel.domainname). Some legacy software uses the domain name of the
host for license verification and other similar functions, so having
this is a very useful feature for people who use such software.

Signed-off-by: Aleksa Sarai <asarai@suse.de>