config: Add Windows Devices to Schema
Signed-off-by: Craig Wilhite <crwilhit@microsoft.com>

This adds Windows Device support to the OCI runtime-spec: code, JSON
schema and markdown documentation. This is a first step in our approach
to lighting up access to host devices for Windows Server containers.

Devices in Windows implement any number of interface class GUIDs. By
configuring a container with a device interface class GUID, we will
plug all devices on the host that implement the interface class GUID
into the device namespace of the container.

While this approach does not allow for the speciication of a single,
particular device, our schema definition is defined to be flexible;
we interpret the 'id' of the device passed in based upon the 'id_type'.