maskedPaths and readonlyPaths: skip unexistent paths
runc ignores unexistent paths in maskedPaths and readonlyPaths. That's
useful for blocking /proc/latency_stats (default in buildah) because
this path is not existing on all kernels.

In this case, no error should be generated. Other errors should be
generated. For example, using readonlyPaths on a unbindable path fails
and this error must be generated, otherwise the path would silently stay
read-write.

Signed-off-by: Alban Crequy <alban@kinvolk.io>