Remove effective and permitted capabilities
Both `effective` and `permitted` capabilities cannot be passed to a
container because the Linux kernel will reset them during the exec of
the pid1 process in the container. In practice, no matter how these
two capabilities are set in a runtime config file, those can be set
in the container to a totally different values.

So let's remove these capabilities from the runtime spec, because it's
not possible to make them work inside containers.

Fixes https://github.com/opencontainers/runtime-spec/issues/979

Signed-off-by: Dongsu Park <dongsu@kinvolk.io>