Change all references from whitelist to allowed
We want to move to more enclusive names/terms in our code, and remove problematic language
from code and comments.

We want to change reference for whitelist/blacklist to allowlist/denylist.

https://www.redhat.com/en/blog/making-open-source-more-inclusive-eradicating-problematic-language

We also want to fix slave references to pty for pseutoterminals.

We will change the slave refererences to whatever the kernel specifies, once the kernel fixes it's
references.

Co-authored-by: Sebastiaan van Stijn <thaJeztah@users.noreply.github.com>
Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>