config-linux: mark memory.kernel[TCP] as NOT RECOMMENDED
Per-cgroup kernel memory accounting (and explicit limiting) is
problematic in the Linux kernel for too many reasons to quote here.
Besides, cgroup v2 does not even have a kernel memory limit knob,
and the one in cgroup v1 is made obsoleted in kernel v5.4 [1].

Mark memory.kernel and memory.kernelTCP as NOT RECOMMENDED, in additon
to OPTIONAL. This is a way to say "we do not anyone (runtimes or users)
to set those limits, unless they have good understanding and strong
reasons to do so".

[1] https://github.com/torvalds/linux/commit/0158115f702b0ba208ab0b

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>