seccomp: Add missing const for seccomp notify action
This wasn't catched before, even though we had working patches for
containerd and runc in advance, as neither containerd nor runc really
use these consts.

In the spec this field is a string[1] and therefore when containerd
parses with json.Unmarshall[2] it works just fine. With runc is not used
either, as it uses a different struct in the libcontainer directory[3].

Therefore, even with patches using this change, this const definition
was missed as it is not used by the patches.

[1]: https://github.com/opencontainers/runtime-spec/blob/a8c4a9ee0f6b5a0b994c5c23c68725394e2b0d9d/specs-go/config.go#L641
[2]: https://github.com/containerd/containerd/blob/8dbe53a2a930af3631229e4d92cf839b64ee5a38/contrib/seccomp/seccomp.go#L36-L40
[3]: https://github.com/opencontainers/runc/pull/2682/files#diff-9915e69bab45a993d366aad4a7d47459d73ec4304b7c33942f197dd221673376R51
[4]: https://github.com/opencontainers/runtime-spec/blob/a8c4a9ee0f6b5a0b994c5c23c68725394e2b0d9d/specs-go/config.go#L614

Signed-off-by: Rodrigo Campos <rodrigo@kinvolk.io>