schema/defs-linux: Fix inconsistencies with seccomp notify
Commit "Add Seccomp Notify support"
(58798e75e9803d99bff5837ff39e9afe2e2efec8) just added
SECCOMP_FILTER_FLAG_NEW_LISTENER to the schema and not to the list of
flags in config-linux.md. However, it was a mistake to add them to the
schema, as the user will never really need to specify that flag.

Signed-off-by: Rodrigo Campos <rodrigo@kinvolk.io>