specs-go/config: add Core Scheduling support
Linux kernel 5.14 adds the support for Core Scheduling.

This allows setting and copying core scheduling 'task cookies' between
the container process and the threads (PID), processes (TGID), and
process groups (PGID), which helps define groups of tasks that can be
co-scheduled on the same core. These groups can be specified either for
security usecases or for performance usecases.

https://github.com/opencontainers/runtime-spec/issues/1113

Signed-off-by: Kailun Qin <kailun.qin@intel.com>