config-linux: add sysDevBlock
This is to expose the /sys/dev/block hierarchy to the container
for devices of mounted filesystems and block devices of the containers.

In rhbz#1772993 the OCP decided to block the system storage information
from unprivileged containers. As a move to re-expose this information
back to the unprivileged containers, this add the option back, leaving
the default behaviour to what the container runtime previously did.

Signed-off-by: Daniel Black <daniel@mariadb.org>