runtime: move prestart hook in lifecycle section
config.md specifies that "`prestart` hooks MUST be called after the
`start` operation is called but before the user-specified program
command is executed." However, the Lifecycle section in runtime.md
contradicts this by showing that `prestart` hooks are called after the
`create` operation is called and before the `start` operation. As the
note in the config.md section for `createRuntime` hooks indicates, the
order of operations in runtime.md#Lifecycle reflects the old runC
behaviour which was contrary to the spec. Update the Lifecycle section
to be congruent with the hook semantics as specified in config.md.

Signed-off-by: Cory Snider <csnider@mirantis.com>